fix : SecurityConfig에 필터 추가
필터 작성 후 적용은 안해서 SecurityContextHolder에 등록되지 않아 유저를 찾을 수 없는 에러 해결

+++ src/main/java/com/gagoo/thiscoding/domain/maria/user/infrastructure/security/JwtUtilImpl.java
@@ -84,7 +84,7 @@ public class JwtUtilImpl implements JwtUtil {
                 .claim(EMAIL, email)
                 .claim(ROLE, role)
                 .issuedAt(new Date(System.currentTimeMillis())) // 발행시간
-                .expiration(new Date(System.currentTimeMillis() + accessTokenExpTime)) // 소멸시간
+                .expiration(new Date(System.currentTimeMillis() + accessTokenExpTime.longValue()))
                 .signWith(secretKey)
                 .compact();
     }
@@ -97,7 +97,7 @@ public class JwtUtilImpl implements JwtUtil {
                 .claim(EMAIL, email)
                 .claim(ROLE, role)
                 .issuedAt(new Date(System.currentTimeMillis())) // 발행시간
-                .expiration(new Date(System.currentTimeMillis() + refreshTokenExpTime)) // 소멸시간
+                .expiration(new Date(System.currentTimeMillis() + refreshTokenExpTime.longValue()))
                 .signWith(secretKey)
                 .compact();
     }
